+++ oshi-core/src/main/java/oshi/software/os/unix/aix/AixFileSystem.java
@@ -16,6 +16,7 @@ import oshi.software.common.AbstractFileSystem;
 import oshi.software.os.OSFileStore;
 import oshi.util.ExecutingCommand;
 import oshi.util.FileSystemUtil;
+import oshi.util.FileUtil;
 import oshi.util.ParseUtil;
 
 /**
@@ -168,5 +169,19 @@ public class AixFileSystem extends AbstractFileSystem {
     @Override
     public long getMaxFileDescriptors() {
         return ParseUtil.parseLongOrDefault(ExecutingCommand.getFirstAnswer("ulimit -n"), 0L);
+    }
+
+    @Override
+    public long getMaxFileDescriptorsPerProcess() {
+        final List<String> lines = FileUtil.readFile("/etc/security/limits");
+        for (final String line : lines) {
+            if (line.startsWith("nofiles(descriptors)")) {
+                final String[] split = line.split("\\w+");
+                if (split.length > 1) {
+                    return ParseUtil.parseLongOrDefault(split[1], Long.MAX_VALUE);
+                }
+            }
+        }
+        return Long.MAX_VALUE;
     }
 }