+++ oshi-core/src/main/java/oshi/software/os/unix/freebsd/FreeBsdFileSystem.java
@@ -146,4 +146,10 @@ public final class FreeBsdFileSystem extends AbstractFileSystem {
     public long getMaxFileDescriptors() {
         return BsdSysctlUtil.sysctl("kern.maxfiles", 0);
     }
+
+    @Override
+    public long getMaxFileDescriptorsPerProcess() {
+        // On FreeBsd there is no process specific system-wide limit, so the general limit is returned
+        return getMaxFileDescriptors();
+    }
 }